+++ stories/templates/validator/component.template.js
@@ -1,5 +1,25 @@
 export const template = `
 <div class="validator-base">
-  <st-input v-model="inputValue" :validator="inputValidator" />
+  <span class="validator-base__caption">Input</span>
+  <div class="validator-base__block">
+    <st-input v-model="inputValue"
+              :validator="inputValidator"
+              placeholder="Input" />
+  </div>
+
+  <span class="validator-base__caption">Select</span>
+  <div class="validator-base__block">
+    <st-select :options="selectOptions"
+               v-model="singleSelectValue"
+               :validator="singleSelectValidator"
+               clearable
+               placeholder="Single Select" />
+    <st-select :options="selectOptions"
+               multiple
+               v-model="multipleSelectValue"
+               :validator="multipleSelectValidator"
+               clearable
+               placeholder="Multiple Select" />
+  </div>
 </div>
 `;