+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.15
+
+### Features
+* Added `Validation` for `Select`
+* Added `NotEmptyArrayRule` rule for `Validation`.
+
+
+
+
 ## 0.2.14
 
 ### Features